be/pyproject.toml
[tool.poetry]
name = "levity-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.1"
python-dateutil = "^2.8.2"
dj-database-url = "^1.3.0"
psycopg2 = "^2.9.5"
retry = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
